+++ packages/generator/generator.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/K-Phoen/grabana/dashboard"
-	"github.com/fusakla/autograf/packages/prometheus"
+	"github.com/fusakla/autograf/packages/model"
+	"github.com/fusakla/sdk"
 )
 
 func newTree(prefix string, level int) *metricsTree {
@@ -19,12 +19,12 @@ func newTree(prefix string, level int) *metricsTree {
 type metricsTree struct {
 	leafs  map[string]*metricsTree
 	prefix string
-	metric *prometheus.Metric
+	metric *model.Metric
 	level  int
 }
 
-func (t *metricsTree) metrics() []*prometheus.Metric {
-	metrics := []*prometheus.Metric{}
+func (t *metricsTree) metrics() []*model.Metric {
+	metrics := []*model.Metric{}
 	if t.metric != nil {
 		metrics = append(metrics, t.metric)
 	}
@@ -38,9 +38,9 @@ func (t *metricsTree) metrics() []*prometheus.Metric {
 	return metrics
 }
 
-func (t *metricsTree) metricGroups() map[string][]*prometheus.Metric {
-	others := []*prometheus.Metric{}
-	groups := map[string][]*prometheus.Metric{}
+func (t *metricsTree) metricGroups() map[string][]*model.Metric {
+	others := []*model.Metric{}
+	groups := map[string][]*model.Metric{}
 	for _, l := range t.leafs {
 		subtreeMetrics := l.metrics()
 		if len(subtreeMetrics) < 3 && t.prefix != "" {
@@ -51,7 +51,9 @@ func (t *metricsTree) metricGroups() map[string][]*prometheus.Metric {
 			}
 		}
 	}
-	if len(others) > 0 {
+	if len(others) == 1 {
+		groups[others[0].Name] = others
+	} else if len(others) > 0 {
 		groups[t.prefix] = others
 	}
 	return groups
@@ -62,7 +64,7 @@ func (t metricsTree) String() string {
 	out := ""
 	out += fmt.Sprintf(indent + t.prefix + "\n")
 	if t.metric != nil {
-		out += fmt.Sprintf(indent + "- " + t.metric.String())
+		out += fmt.Sprintf("%s- %v", indent, t.metric)
 	}
 	for _, l := range t.leafs {
 		out += "\n" + l.String()
@@ -70,10 +72,10 @@ func (t metricsTree) String() string {
 	return out
 }
 
-func (t *metricsTree) add(metric prometheus.Metric) {
+func (t *metricsTree) add(metric *model.Metric) {
 	strippedMetricName := strings.TrimPrefix(strings.TrimPrefix(metric.Name, t.prefix), "_")
 	if strippedMetricName == "" {
-		t.metric = &metric
+		t.metric = metric
 		return
 	}
 	parts := strings.SplitN(strippedMetricName, "_", 2)
@@ -84,19 +86,32 @@ func (t *metricsTree) add(metric prometheus.Metric) {
 	t.leafs[prefix].add(metric)
 }
 
-func groupMetrics(metrics map[string]prometheus.Metric) map[string][]*prometheus.Metric {
+func groupMetrics(metrics map[string]*model.Metric) map[string][]*model.Metric {
 	tree := newTree("", 0)
+	metricGroups := map[string][]*model.Metric{}
 	for _, m := range metrics {
+		if m.Config.Row != "" {
+			r, ok := metricGroups[m.Config.Row]
+			if !ok {
+				metricGroups[m.Config.Row] = []*model.Metric{m}
+			} else {
+				metricGroups[m.Config.Row] = append(r, m)
+			}
+			continue
+		}
 		tree.add(m)
 	}
-	return tree.metricGroups()
+	for k, v := range tree.metricGroups() {
+		metricGroups[k] = v
+	}
+	return metricGroups
 }
 
-func Generate(name, datasource, selector string, filerVariables []string, metrics map[string]prometheus.Metric) (*dashboard.Builder, error) {
-	metricGroups := groupMetrics(metrics)
+func Generate(name, datasource, selector string, filerVariables []string, metrics map[string]*model.Metric) (*sdk.Board, error) {
+	metricGroups := groupMetrics(metrics)	
 	dashboard, err := newDashboard(name, datasource, selector, filerVariables, metricGroups)
 	if err != nil {
 		return nil, err
 	}
-	return &dashboard, nil
+	return dashboard, nil
 }