+++ packages/prometheus/api_fetcher.go
@@ -5,14 +5,22 @@ import (
 	"fmt"
 	"net/http"
 	"regexp"
-	"strings"
 	"time"
 
+	autograf_model "github.com/fusakla/autograf/packages/model"
 	"github.com/prometheus/client_golang/api"
 	v1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	"github.com/prometheus/common/model"
 	"github.com/sirupsen/logrus"
 )
+
+var (
+	specialSuffixRegexp = regexp.MustCompile(`(.+)_(total|info|sum|count|bucket)`)
+)
+
+func stripSpecialSuffixes(metricName string) string {
+	return specialSuffixRegexp.ReplaceAllString(metricName, "$1")
+}
 
 func NewClient(logger logrus.FieldLogger, prometheusURL string, transport http.RoundTripper) (*Client, error) {
 	cfg := api.Config{Address: prometheusURL}
@@ -32,14 +40,6 @@ func NewClient(logger logrus.FieldLogger, prometheusURL string, transport http.R
 type Client struct {
 	v1api  v1.API
 	logger logrus.FieldLogger
-}
-
-var (
-	specialSuffixRegexp = regexp.MustCompile(`(.+)_(total|info|sum|count|bucket)`)
-)
-
-func stripSpecialSuffixes(metricName string) string {
-	return specialSuffixRegexp.ReplaceAllString(metricName, "$1")
 }
 
 func (c *Client) queryMetricsMetadata(ctx context.Context) (map[string][]v1.Metadata, error) {
@@ -69,28 +69,7 @@ func (c *Client) queryMetricsForSelector(ctx context.Context, selector string) (
 	}
 }
 
-func dropCreatedMetrics(metrics map[string]Metric) {
-	for k := range metrics {
-		metricName := strings.TrimSuffix(k, "_created")
-		if k == metricName {
-			continue
-		}
-		if _, ok := metrics[metricName]; ok {
-			delete(metrics, k)
-			continue
-		}
-		if _, ok := metrics[metricName+"_total"]; ok {
-			delete(metrics, k)
-			continue
-		}
-		if _, ok := metrics[metricName+"_count"]; ok {
-			delete(metrics, k)
-			continue
-		}
-	}
-}
-
-func (c *Client) MetricsForSelector(ctx context.Context, selector string) (map[string]Metric, error) {
+func (c *Client) MetricsForSelector(ctx context.Context, selector string) (map[string]*autograf_model.Metric, error) {
 	samples, err := c.queryMetricsForSelector(ctx, selector)
 	if err != nil {
 		return nil, err
@@ -99,37 +78,25 @@ func (c *Client) MetricsForSelector(ctx context.Context, selector string) (map[s
 	if err != nil {
 		return nil, err
 	}
-	metrics := map[string]Metric{}
+	metrics := map[string]*autograf_model.Metric{}
 	for _, s := range samples {
 		metricName := string(s.Metric["__name__"])
-		metricMetadata, ok := metadata[metricName]
+		metricMetadata, ok := metadata[stripSpecialSuffixes(metricName)]
 		if !ok {
-			metricMetadata, ok = metadata[stripSpecialSuffixes(metricName)]
-			if !ok {
-				metricMetadata = []v1.Metadata{}
+			metricMetadata = metadata[metricName]
+		}
+		if len(metricMetadata) > 0 {
+			metrics[metricName] = &autograf_model.Metric{
+				Name:       metricName,
+				MetricType: autograf_model.MetricType(metricMetadata[0].Type),
+				Help:       metricMetadata[0].Help,
+				Unit:       autograf_model.MetricUnit(metricMetadata[0].Unit),
+			}
+		} else {
+			metrics[metricName] = &autograf_model.Metric{
+				Name: metricName,
 			}
 		}
-		if len(metricMetadata) == 0 {
-			metricMetadata = guessMetricMetadata(metricName)
-			c.logger.WithField("metric_name", metricName).WithField("metadata", metricMetadata).Warn("failed to get metric metadata, trying to guess...")
-		}
-		m := Metric{
-			Name:       metricName,
-			MetricType: MetricType(metricMetadata[0].Type),
-			Help:       metricMetadata[0].Help,
-			Unit:       MetricUnit(metricMetadata[0].Unit),
-		}
-		if m.Unit == "" {
-			m.Unit = guessMetricUnit(metricName)
-		}
-		if m.MetricType == MetricTypeHistogram && (strings.HasSuffix(string(m.Name), "_sum") || strings.HasSuffix(string(m.Name), "_count")) {
-			m.MetricType = MetricTypeCounter
-		}
-		if strings.HasSuffix(m.Name, "_info") {
-			m.MetricType = MetricTypeInfo
-		}
-		metrics[metricName] = m
 	}
-	dropCreatedMetrics(metrics)
 	return metrics, nil
 }